refactor: projectsTab [ADR] Projects Tab Design & Implementation #6
